+++ user/src/main/java/silverpotion/userserver/report/controller/ReportController.java
@@ -14,7 +14,7 @@ import silverpotion.userserver.report.service.ReportService;
 
 
 @RestController
-@RequestMapping("/silverpotion/admins/report")
+@RequestMapping("/silverpotion/report")
 public class ReportController {
     private final ReportService reportService;
 
@@ -29,27 +29,5 @@ public class ReportController {
         return new ResponseEntity<>(new CommonDto(HttpStatus.CREATED.value(),"신고 등록 완료",report),HttpStatus.CREATED);
     }
 
-//    신고 목록 조회
-    @GetMapping("/list")
-    @PreAuthorize("hasRole('ADMIN')")
-    public ResponseEntity<?> getReportList(@RequestHeader("X-User-LoginId") String loginId, Pageable pageable, ReportRequestDto reportRequestDto){
-        Page<ReportResponseDto> reports = reportService.findAllReports(loginId, pageable,reportRequestDto);
-        return new ResponseEntity<>(new CommonDto(HttpStatus.OK.value(),"신고 유저 조회 성공",reports),HttpStatus.OK);
-    }
 
-//    신고 상세 조회
-    @GetMapping("/detail/{reportId}")
-    @PreAuthorize("hasRole('ADMIN')")
-    public ResponseEntity<?> getReportDetails(@PathVariable Long reportId){
-        ReportDetailListDto dto = reportService.getReportDetails(reportId);
-        return new ResponseEntity<>(new CommonDto(HttpStatus.OK.value(), "detail is uploaded successfully",dto),HttpStatus.OK);
-    }
-
-//    특정 신고 처리
-    @PostMapping("{reportId}/process")
-    @PreAuthorize("hasRole('ADMIN')")
-    public ResponseEntity<?> getReportProcess(@PathVariable Long reportId, @RequestBody @Valid ReportProcessResDto dto){
-        ReportProcessResDto reportProcessResDto = reportService.processReport(reportId,dto);
-        return new ResponseEntity<>(new CommonDto(HttpStatus.OK.value(), "process is uploaded successfully",reportProcessResDto),HttpStatus.OK);
-    }
 }
